fixed util renaming
